Split Directory Search as a base for further changes
